cloud_orchestrator/requirements.txt
python-keystoneclient
python-glanceclient
python-neutronclient
python-novaclient

